1.0

+++ Ship.h
@@ -6,7 +6,7 @@
 class Ship : public Item{
 public:
   int health;
-  size_t bullet, missile, torpedoes;
+  int bullet, missile, torpedoes;
   Ship(Game* g, size_t r, size_t c, size_t v,size_t t):Item(g,r,c,v,t){}
   virtual ~Ship(){}
   virtual void move(size_t direction, size_t step);
